# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clinic",
        "description": "API پروژه ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/appointments": {
            "post": {
                "tags": [
                    "Appointment"
                ],
                "summary": "رزرو نوبت",
                "description": "رزرو یک نوبت موجود توسط کاربر لاگین‌شده",
                "operationId": "ReserveAppointment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "appointment_id"
                                ],
                                "properties": {
                                    "appointment_id": {
                                        "description": "شناسه نوبت",
                                        "type": "integer",
                                        "example": 12
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "رزرو با موفقیت انجام شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 12
                                        },
                                        "provider_id": {
                                            "type": "integer",
                                            "example": "دکتر محمد صادقی"
                                        },
                                        "customer_id": {
                                            "type": "integer",
                                            "example": "مهدی محمدی"
                                        },
                                        "appointment_date": {
                                            "type": "string",
                                            "example": "1404-10-13"
                                        },
                                        "start_time": {
                                            "type": "string",
                                            "example": "09:00"
                                        },
                                        "end_time": {
                                            "type": "string",
                                            "example": "09:30"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "reserved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "نوبت قبلاً رزرو شده یا در حال رزرو است",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "قبلا رزرو شده است"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "کاربر احراز هویت نشده است"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/appointments/{id}": {
            "put": {
                "tags": [
                    "Appointment"
                ],
                "summary": "تغییر نوبت رزرو شده",
                "description": "تغییر نوبت فعلی کاربر به نوبت جدید در صورت خالی بودن",
                "operationId": "ChangeReservation",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "شناسه نوبت فعلی کاربر",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "appointment_id"
                                ],
                                "properties": {
                                    "appointment_id": {
                                        "description": "شناسه نوبت جدید برای تغییر",
                                        "type": "integer",
                                        "example": 25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "نوبت با موفقیت تغییر کرد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "provider_id": {
                                            "type": "integer",
                                            "example": "دکتر محمد صادقی"
                                        },
                                        "customer_id": {
                                            "type": "integer",
                                            "example": "علی رضایی"
                                        },
                                        "appointment_date": {
                                            "type": "string",
                                            "example": "1404-10-13"
                                        },
                                        "start_time": {
                                            "type": "string",
                                            "example": "11:00"
                                        },
                                        "end_time": {
                                            "type": "string",
                                            "example": "11:30"
                                        },
                                        "status": {
                                            "type": "string",
                                            "example": "reserved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "نوبت جدید قبلاً رزرو شده است",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "نوبت مورد نظر قبلا رزرو شده است"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "کاربر احراز هویت نشده است"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/appointments/my": {
            "get": {
                "tags": [
                    "Appointment"
                ],
                "summary": "لیست نوبت‌های من",
                "description": "دریافت لیست نوبت‌های رزرو شده توسط کاربر لاگین‌شده",
                "operationId": "MyReservations",
                "responses": {
                    "200": {
                        "description": "لیست نوبت‌ها با موفقیت دریافت شد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 12
                                            },
                                            "provider_name": {
                                                "type": "string",
                                                "example": "دکتر محمد صادقی"
                                            },
                                            "appointment_date": {
                                                "type": "string",
                                                "example": "1404-10-14"
                                            },
                                            "start_time": {
                                                "type": "string",
                                                "example": "10:00"
                                            },
                                            "end_time": {
                                                "type": "string",
                                                "example": "10:30"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "reserved"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "کاربر احراز هویت نشده است"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "Bearer {token}"
            }
        }
    }
}